Copy all available bootloaders
Since bootloaders being available don't cause a conflict and the desired
bootloader is chosen at compile time, just add all available bootloaders
into the test folder. This eliminates an issue where the selected
bootloader was not first in the if else chain.

This probably should be fixed by having the configured bootloader also
set in the test-init.sh and the script would fail when that bootloader
wasn't available to avoid debugging issues but for now this was easier
to implement.