Begin deduplication to enable syslinux
The kernel blob management is a standard mechanism employed by CBM itself,
thus the bootloader implementation should not be responsible for the propogation
of these files.

Using this new model, CBM itself will first attempt safe manipulation of the
kernel assets, i.e the namespaced kernel blob, and only when that has completed
successfully will we hand over to the bootloader implementation.

This also ensures that when implementing new bootloaders there is no need
for duplicated kernel blob management logic, or a scenario in which they could
compete with one another to thrash the boot partition.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>