Handle installation of syslinux configs
Unlike other bootloader implementations so far, syslinux has a single
config for all kernels. To handle this case install simply adds the
kernel to a static array of kernels that will be installed once the call
to boot_manager_set_default_kernel is placed in the update path.

Remove is a no op in this case as well since the only kernels that will
be present in the config will be those that are expliticly installed
during the update loop.