Add the new logging functionality
These new logging functions will be used in an upcoming commit as the sole
form of stderr messaging from clr-boot-manager. Unlike ordinary logging
mechanisms, this should be used sparingly.

It is expected that consumers of the API do not directly use the cbm_log
function, rather they use the leveled macros, i.e. LOG_DEBUG. The default
log level is CBM_LOG_ERROR, matching the current functionality.

In all instances, any log message with a level below CBM_LOG_ERROR will not
be emitted, *unless* it is overriden with the environmental variable CBM_DEBUG.
This should be set to the value of the desired log level correlating with the
CbmLogLevel enum. Thus, setting CBM_DEBUG=1 will set the minimum debug level
to CBM_LOG_DEBUG, enabling all DEBUG and higher messages to be emitted.

The main reason for this change is to enable deeper inspection of the update
process to help triage any current and future issues, specifically during
image generation and update automation. With this change, and the next comit
to introduce application of the log functionality, devops will be able to
run image builds with CBM_DEBUG set in order to create more useful log output.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>