bootman: Don't double-probe our prefix
Currently we're attempting to _always_ probe "/", and then potentially later
we update to a given prefix in set_prefix, meaning 2 sets of log outputs,
potentially invalid content, and checks for UEFI vs legacy in a premature
fashion.

As a result of this change, the CLI frontend now explicitly sets the prefix
in all instances.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>