update: Allow upgrades in the absence of a known, running kernel
This ensures that in the case of errors, where the running kernel cannot be
reliably determined, we still update past that issue, with the exception of
disabling garbage collection during this update cycle.

It is expected that there will be updates to fix this temporary issue, and
that cleaning would then happen on the update after that.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>