bootman: Add fallback kernel detection for broken versions
In some cases we may find that uname doesn't quite match the version number,
but all other fields are matching. This can leave abandoned kernels on disk
during update that are never collected, and even prevent updates. As such
we do a type and release comparison, which would be unique to their type within
the shipped kernels.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>