bootman: Compare stream buffer in cbm_files_match
Instead of computing the hash sum for both of the files, we simply open both
of the files, and then use memcmp to compare them on a byte-by-byte basis.
If the lengths don't match, or we failed to open one of the files, we simply
return false because one of them indeed does not match.

This preserves the behaviour of the previous cbm_files_match function whilst
vastly simplifying it, and is arguably faster than using cryptographic
hashes for a trivial byte stream comparison.

This in turn will help us to remove the now unnecessary OpenSSL dependency
from clr-boot-manager, protecting against future ABI changes in the OpenSSL
project (1.1.0), vital for a pain-free upgrade process for such a critical
component.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>