timeout: Change configuration to /etc/kernel/timeout
Recent changes have made CBM read files from /etc/kernel, such as the
cmdline. Keeping in line with this simpler approach, i.e. "echo >" and
upgrade, we use a simpler filename.

To enhance ease of use, we'll also create the /etc/kernel directory if
it doesn't already exist, so that there are less steps involved for the
user.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>