bootman: Don't hardcode os-name and vendor-prefix values
The vendor prefix options is now taken from a configure time option, which
ensures that it remains constant over the lifetime of a distribution, and
not influenced by other factors such as user changes.

The "OS name" field is now taken from the PRETTY_NAME field in the
os-release file if we find it. If we don't encounter this file, or cannot
parse it, we simply fallback to using the builtin PRETTY_NAME of
"generic-linux-os".

This allows a more dynamic approach to the "branding" of the boot entries,
and allows pushing PRETTY_NAME changes including the version without having
to push a larger update delta including clr-boot-manager itself.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>