bootman: Add support for installation and removal of initrd files
This change adds preliminary support for initial ramdisk files in
distributions that make use of them. It is expected that the initrd
file suffix match the kernel path, and be prefixed with "initrd-".

This should be shipped within the kernel packaging or deployment
mechanism in the distribution. As such it removes a great deal of
scope for error, by ensuring the initrd's are sane at source, and
not rely on initrd generation on the installation (i.e. via use of
dracut.)

As users may actually have requirements for overwriting their initramfs,
we'll first look for the file in /etc/kernel and use that, otherwise we
will fallback to the distribution provided file.

Upon removal, the user's initrd is left intact, and we only remove the
files within our own domain.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>