bootloaders/systemd: Add support for CbmWriter and initrd
This change helps to simplify the per-kernel config writer by switching
to the CbmWriter interface, allowing us to reduce the amount of times
we check for string/alloc failures.

Additionally, we can now very simply plug in initrd support by simply
emitting a new cbm_writer_append_printf line if the initrd file is known
for the given kernel.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>