bootman: Move the files implementation into lib/
The majority of the "generic" library code for clr-boot-manager now exists
inside lib/. The files code has expanded much since the first introduction
and has become the generic stopping ground for "nice API" code.

To address the logical organisation issue, we make this part of the codebase
compliant with the rest, to enforce a proper domain approach.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>