tests: Restore the auto-update tests
These tests verify the basic install & update functionality for the
bootloaders, not the kernels. It should be noted that in the case of
syslinux, we have a restricted feature-set. As such we *always* report
that syslinux needs an update when the source file changes, because
it's far cheaper than actually reading the block device itself.

This change also addresses the gptmbr bin file being 1 byte too long,
which was revealed during the development of this test.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>