tests: Restore test for UEFI bootloader removal
This test ensures that the UEFI-specific bootloader removal functionality
does as advertised, and truly does remove all of our "bits" from the
target system when requested.

This test is distinctly more messy than the current tests as it has to
know exactly which files should be gone due to the opaque bootloader
system.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>