bootman: Always reinit using a valid boot directory
Previously the modify_bootloader invocation would attempt to reinit itself
with the abs_bootdir. However, in the instance of a native image, we've
had no reason to set a new boot_dir, thus this value is now NULL, leading
to set_boot_dir to fail for the first time.

Once this is set here, i.e. because we're looking at a real root, we
fire off the reinspection and everything "just works".

This change helps, in part, issue #54.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>